Extract native job status assembly from submissions.ts
Moves nativeJobStatus into delivery/native-job-status.ts as
createNativeJobStatusAssembler, a factory over store/gamesStore/
managedAvailabilityGate/now/builderOf. Continues the Wave B seam
extraction from submissions.ts (Phase 3 decomposition).
